+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 TBD
+* Adding Get Device Token API for shared device mode #2980
 * Improve UI tests performance #2981
 
 ## [2.11.0]: